scripts/requirements.txt
# Lab Generator Dependencies
# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
lxml>=5.0.0

# Google Gemini API (new package)
google-genai>=1.0.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
jinja2>=3.1.0

# Rate limiting and retries
tenacity>=8.2.0
